Get rid of utils.go
